+++ my-app/src/SignIn.css
@@ -1,0 +1,79 @@
+/* SignIn.css */
+
+/* Sign-in container with shadow */
+.signin-container {
+    max-width: 400px;
+    margin: 50px auto;
+    padding: 40px 20px;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2); /* Add shadow to the form */
+    background-color: #fff;
+    min-height: 500px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  /* Enlarging and styling the Sign Up text */
+  p a {
+    font-size: 18px; /* Increase font size for 'Sign Up' */
+    font-weight: bold; /* Make it bold */
+    color: #007bff; /* Match button/link color */
+    text-decoration: none; /* Remove underline */
+  }
+  
+  p a:hover {
+    text-decoration: underline; /* Add underline on hover */
+  }
+  
+  /* Shadows for inputs */
+  input {
+    padding: 8px;
+    font-size: 14px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    width: 95%;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Shadow for inputs */
+    transition: box-shadow 0.3s ease; /* Smooth shadow transition */
+  }
+  
+  input:focus {
+    outline: none;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2); /* Stronger shadow on focus */
+  }
+  
+  /* Style for the button */
+  button {
+    padding: 10px 15px;
+    background-color: #007bff;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s ease; /* Smooth hover transition */
+  }
+  
+  button:hover {
+    background-color: #0056b3;
+  }
+  
+  /* Error styling */
+  .error {
+    color: red;
+    font-size: 14px;
+    margin-bottom: 10px;
+    text-align: center;
+  }
+  
+  /* Add spacing between the form-group elements */
+  .form-group {
+    margin-bottom: 20px; /* Increase space between inputs */
+  }
+  
+  /* Specific spacing for the password input field */
+  #password {
+    margin-bottom: 30px; /* Add extra space after the password field */
+  }
+
