+++ my-app/src/TransferPayments.js
@@ -4,15 +4,20 @@ import './TransferPayments.css'; // Optional: Add a separate CSS file for this p
 
 const TransferPayments = ({ balances, onTransfer }) => {
   const [amount, setAmount] = useState(0);
-  const [selectedAccount, setSelectedAccount] = useState("checking"); // Default to checking account
+  const [sourceAccount, setSourceAccount] = useState("checking"); // Default source account
+  const [targetAccount, setTargetAccount] = useState("savings"); // Default target account
   const navigate = useNavigate(); // Initialize useNavigate hook
 
   const handleAmountChange = (e) => {
     setAmount(parseFloat(e.target.value)); // Set the amount from the input
   };
 
-  const handleAccountChange = (e) => {
-    setSelectedAccount(e.target.value); // Set selected account
+  const handleSourceAccountChange = (e) => {
+    setSourceAccount(e.target.value); // Set source account
+  };
+
+  const handleTargetAccountChange = (e) => {
+    setTargetAccount(e.target.value); // Set target account
   };
 
   const handleSubmit = (e) => {
@@ -21,50 +26,63 @@ const TransferPayments = ({ balances, onTransfer }) => {
       alert("Please enter a valid amount");
       return;
     }
-    // Call the onTransfer function passed down as prop
-    onTransfer(selectedAccount, amount);
+    if (balances[sourceAccount] < amount) {
+      alert("Insufficient funds in the source account");
+      return;
+    }
+
+    // Perform the transfer from source account to target account
+    onTransfer(sourceAccount, targetAccount, amount);
     navigate("/account"); // Navigate to Account page after transfer
   };
 
   return (
     <div className="transferContainer">
       <div className="transfer-payments-container">
-      <h1>Transfer & Payments</h1>
-      <p>Here you can transfer money and make payments.</p>
+        <h1>Transfer & Payments</h1>
+        <p>Here you can transfer money and make payments.</p>
 
-      <div className="transfer-form">
-        <h3>Transfer Funds</h3>
-        <form onSubmit={handleSubmit}>
-          <label>
-            Amount:
-            <input
-              type="number"
-              name="amount"
-              value={amount}
-              onChange={handleAmountChange}
-            />
-          </label>
+        <div className="transfer-form">
+          <h3>Transfer Funds</h3>
+          <form onSubmit={handleSubmit}>
+            <label>
+              Amount:
+              <input
+                type="number"
+                name="amount"
+                value={amount}
+                onChange={handleAmountChange}
+              />
+            </label>
 
-          <label>
-            Select Account:
-            <select value={selectedAccount} onChange={handleAccountChange}>
-              <option value="checking">Checking Account</option>
-              <option value="savings">Savings Account</option>
-              <option value="creditCard">Credit Card</option>
-            </select>
-          </label>
+            <label>
+              Source Account:
+              <select value={sourceAccount} onChange={handleSourceAccountChange}>
+                <option value="checking">Checking Account</option>
+                <option value="savings">Savings Account</option>
+                <option value="creditCard">Credit Card</option>
+              </select>
+            </label>
 
-          <button type="submit">Transfer</button>
-        </form>
+            <label>
+              Target Account:
+              <select value={targetAccount} onChange={handleTargetAccountChange}>
+                <option value="checking">Checking Account</option>
+                <option value="savings">Savings Account</option>
+                <option value="creditCard">Credit Card</option>
+              </select>
+            </label>
+
+            <button type="submit">Transfer</button>
+          </form>
+        </div>
+
+        {/* Back to Dashboard button */}
+        <button onClick={() => navigate("/dashboard")} className="back-button">
+          Back to Dashboard
+        </button>
       </div>
-
-      {/* Back to Dashboard button */}
-      <button onClick={() => navigate("/dashboard")} className="back-button">
-        Back to Dashboard
-      </button>
     </div>
-    </div>
-    
   );
 };
